Proper functional alpha version.
